[fix] Microdecision : do not miserably crash if the lock does not exist when attempting to release it

+++ moulinette/core.py
@@ -497,7 +497,10 @@ class MoulinetteLock(object):
 
         """
         if self._locked:
-            os.unlink(self._lockfile)
+            if os.path.exists(self._lockfile):
+                os.unlink(self._lockfile)
+            else:
+                logger.warning("Uhoh, somehow the lock %s did not exist ..." % self._lockfile)
             logger.debug('lock has been released')
             self._locked = False
